# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aplikasi Gacor API Documentation",
        "description": "Documentation for the Aplikasi Gacor Package Delivery Management System",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": ""
        }
    ],
    "paths": {
...
        "/api/penerimas": {
            "get": {
                "tags": [
                    "Penerima"
                ],
                "summary": "Ambil semua data penerima",
                "operationId": "f0cda67bae09ed34db42440b474e8c43",
                "responses": {
                    "200": {
                        "description": "Data penerima berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Penerima"
                ],
                "summary": "Buat penerima baru",
                "operationId": "e6c756465860b9d30a6d7182222c1dd3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama",
                                    "alamat",
                                    "nomor_telepon"
                                ],
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "John Doe"
                                    },
                                    "alamat": {
                                        "type": "string",
                                        "example": "Jl. Sudirman No. 123, Jakarta"
                                    },
                                    "nomor_telepon": {
                                        "type": "string",
                                        "maxLength": 15,
                                        "example": "081234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Penerima berhasil dibuat"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/penerimas/{id}": {
            "get": {
                "tags": [
                    "Penerima"
                ],
                "summary": "Ambil detail penerima",
                "operationId": "5054da6c3af46a8685f62cf53db4b435",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Penerima ditemukan"
                    },
                    "404": {
                        "description": "Penerima tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Penerima"
                ],
                "summary": "Update data penerima",
                "operationId": "7a6667025f8cd147ea4be7ec7896a155",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Jane Doe"
                                    },
                                    "alamat": {
                                        "type": "string",
                                        "example": "Jl. Thamrin No. 456, Jakarta"
                                    },
                                    "nomor_telepon": {
                                        "type": "string",
                                        "maxLength": 15,
                                        "example": "081234567891"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Penerima berhasil diperbarui"
                    },
                    "404": {
                        "description": "Penerima tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Penerima"
                ],
                "summary": "Hapus penerima",
                "operationId": "89e17ad88c30c949d5b5ca0fe2535abb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Penerima berhasil dihapus"
                    },
                    "404": {
                        "description": "Penerima tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/penerimas/filter/nama": {
            "get": {
                "tags": [
                    "Penerima"
                ],
                "summary": "Filter penerima berdasarkan nama (partial match)",
                "operationId": "8bb488ab9de6805015f4e6f2aee23a3b",
                "parameters": [
                    {
                        "name": "nama",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data penerima yang difilter berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/penerimas/filter/nomor_telepon": {
            "get": {
                "tags": [
                    "Penerima"
                ],
                "summary": "Filter penerima berdasarkan nomor telepon (exact match)",
                "operationId": "a8266953ba71cb9598637a7cbc7a85c8",
                "parameters": [
                    {
                        "name": "nomor_telepon",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data penerima yang difilter berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}